refactor<CommentController>:댓글, 대댓글관련 컨트롤러, 서비스 공통 로직 추상화, pk -> id로 변경

+++ core/src/main/java/com/wemingle/core/domain/comment/service/ReplyService.java
@@ -7,9 +7,12 @@ import com.wemingle.core.domain.comment.repository.CommentRepository;
 import com.wemingle.core.domain.comment.repository.ReplyRepository;
 import com.wemingle.core.domain.img.service.S3ImgService;
 import com.wemingle.core.domain.post.entity.TeamPost;
+import com.wemingle.core.domain.post.service.TeamPostService;
 import com.wemingle.core.domain.team.entity.Team;
 import com.wemingle.core.domain.team.entity.TeamMember;
 import com.wemingle.core.domain.team.repository.TeamMemberRepository;
+import com.wemingle.core.domain.team.service.TeamMemberService;
+import com.wemingle.core.global.exception.NotWriterException;
 import com.wemingle.core.global.exceptionmessage.ExceptionMessage;
 import com.wemingle.core.global.util.CommentResponseUtil;
 import jakarta.persistence.EntityNotFoundException;
@@ -28,19 +31,19 @@ import java.util.List;
 @Transactional(readOnly = true)
 public class ReplyService {
     private final CommentRepository commentRepository;
-    private final TeamMemberRepository teamMemberRepository;
+    private final TeamPostService teamPostService;
     private final ReplyRepository replyRepository;
     private final S3ImgService s3ImgService;
+    private final TeamMemberService teamMemberService;
 
     @Value("${wemingle.ip}")
     private String serverIp;
 
     @Transactional
     public void saveReply(ReplyDto.RequestReplySave saveDto, String memberId){
-        Comment comment = commentRepository.findByIdFetchPost(saveDto.getCommentPk())
+        Comment comment = commentRepository.findByIdFetchPost(saveDto.getCommentId())
                 .orElseThrow(() -> new EntityNotFoundException(ExceptionMessage.COMMENT_NOT_FOUND.getExceptionMessage()));
-        TeamMember requester = teamMemberRepository.findByTeamAndMember_MemberId(comment.getTeamPost().getTeam(), memberId)
-                .orElseThrow(() -> new EntityNotFoundException(ExceptionMessage.TEAM_MEMBER_NOT_FOUND.getExceptionMessage()));
+        TeamMember requester = teamMemberService.findByTeamAndMember_MemberId(comment.getTeamPost().getTeam(), memberId);
 
         replyRepository.save(Reply.builder()
                 .comment(comment)
@@ -52,40 +55,42 @@ public class ReplyService {
     }
 
     public boolean isReplyWriter(Long replyPk, String memberId){
-        Reply reply = replyRepository.findById(replyPk)
-                .orElseThrow(() -> new EntityNotFoundException(ExceptionMessage.REPLY_NOT_FOUND.getExceptionMessage()));
-        TeamMember requester = teamMemberRepository.findByTeamAndMember_MemberId(reply.getWriter().getTeam(), memberId)
-                .orElseThrow(() -> new EntityNotFoundException(ExceptionMessage.TEAM_MEMBER_NOT_FOUND.getExceptionMessage()));
+        Reply reply = findById(replyPk);
+        TeamMember requester = teamMemberService.findByTeamAndMember_MemberId(reply.getWriter().getTeam(), memberId);
 
         return reply.isWriter(requester);
     }
 
     @Transactional
-    public void updateReply(ReplyDto.RequestReplyUpdate updateDto){
-        Reply reply = replyRepository.findById(updateDto.getReplyPk())
-                .orElseThrow(() -> new EntityNotFoundException(ExceptionMessage.REPLY_NOT_FOUND.getExceptionMessage()));
+    public void updateReply(ReplyDto.RequestReplyUpdate updateDto, String memberId) {
+        if (!isReplyWriter(updateDto.getReplyId(), memberId)){
+            throw new NotWriterException();
+        }
+        Reply reply = findById(updateDto.getReplyId());
 
         reply.updateContent(updateDto.getContent());
     }
 
     @Transactional
-    public void deleteReply(ReplyDto.RequestReplyDelete deleteDto){
-        Reply reply = replyRepository.findById(deleteDto.getReplyPk())
-                .orElseThrow(() -> new EntityNotFoundException(ExceptionMessage.REPLY_NOT_FOUND.getExceptionMessage()));
-        TeamPost teamPost = replyRepository.findTeamPostByReplyPk(deleteDto.getReplyPk())
-                .orElseThrow(() -> new EntityNotFoundException(ExceptionMessage.POST_NOT_FOUND.getExceptionMessage()));
+    public void deleteReply(ReplyDto.RequestReplyDelete deleteDto, String memberId) {
+        if (!isReplyWriter(deleteDto.getReplyId(), memberId)){
+            throw new NotWriterException();
+        }
+
+        Reply reply = findById(deleteDto.getReplyId());
+        TeamPost teamPost = teamPostService.findById(deleteDto.getTeamPostId());
+
         reply.delete();
         teamPost.reduceReplyCnt();
     }
 
-    public ReplyDto.ResponseRepliesRetrieve getReplies(Long nextIdx, Long commentPk, String memberId){
-        List<Reply> replies = replyRepository.findRepliesByNextIdx(nextIdx, commentPk);
-        Team team = commentRepository.findTeam(commentPk)
+    public ReplyDto.ResponseRepliesRetrieve getReplies(Long nextIdx, Long commentId, String memberId){
+        List<Reply> replies = replyRepository.findRepliesByNextIdx(nextIdx, commentId);
+        Team team = commentRepository.findTeam(commentId)
                 .orElseThrow(() -> new EntityNotFoundException(ExceptionMessage.TEAM_NOT_FOUND.getExceptionMessage()));
-        TeamMember requester = teamMemberRepository.findByTeamAndMember_MemberId(team, memberId)
-                .orElseThrow(() -> new EntityNotFoundException(ExceptionMessage.TEAM_MEMBER_NOT_FOUND.getExceptionMessage()));
+        TeamMember requester = teamMemberService.findByTeamAndMember_MemberId(team, memberId);
 
-        return createResponseReplies(commentPk, replies, requester);
+        return createResponseReplies(commentId, replies, requester);
     }
 
     public ReplyDto.ResponseRepliesRetrieve createResponseReplies(Long commentPk, List<Reply> replies, TeamMember requester) {
@@ -112,5 +117,10 @@ public class ReplyService {
 
     public List<Reply> findRankRepliesByComments(List<Long> commentIds){
         return replyRepository.findRankRepliesByComments(commentIds);
+    }
+
+    public Reply findById(Long replyId) {
+        return replyRepository.findById(replyId)
+                .orElseThrow(() -> new EntityNotFoundException(ExceptionMessage.REPLY_NOT_FOUND.getExceptionMessage()));
     }
 }
